Finish moving all the PostIcons over
