+++ src/components/assets/Icons.js
@@ -204,6 +204,13 @@ export const GridIcon = () =>
     </g>
   </SVGIcon>
 
+export const HeartIcon = () =>
+  <SVGIcon className="HeartIcon">
+    <g>
+      <path d="M10,7.4c0-1.8,1.5-3.2,3.3-3.2s3.3,1.4,3.3,3.2c0,4.5-6.5,8.4-6.5,8.4S3.5,12,3.5,7.4c0-1.8,1.5-3.2,3.3-3.2S10,5.6,10,7.4z" />
+    </g>
+  </SVGIcon>
+
 export const ListIcon = () =>
   <SVGIcon className="ListIcon">
     <g>
@@ -215,6 +222,16 @@ export const ListIcon = () =>
 export const MarkerIcon = () =>
   <SVGIcon className="MarkerIcon">
     <path className="svg-fill" d="M10,2C6.7,2,4,4.7,4,8c0,3.7,4.3,9.4,5.6,11.1c0.2,0.3,0.6,0.3,0.8,0C11.7,17.4,16,11.8,16,8 C16,4.7,13.3,2,10,2z M10,9.9C9,9.9,8.1,9,8.1,8C8.1,7,9,6.1,10,6.1S11.9,7,11.9,8C11.9,9,11,9.9,10,9.9z" />
+  </SVGIcon>
+
+export const PencilIcon = () =>
+  <SVGIcon className="PencilIcon">
+    <g>
+      <polygon points="12.6,4 16,7.3 7.4,15.9 4,15.9 4,12.5" />
+    </g>
+    <g>
+      <line x1="10.2" y1="6.4" x2="13.6" y2="9.7" />
+    </g>
   </SVGIcon>
 
 export const PlusIconHeader = () =>
@@ -236,6 +253,28 @@ export const PlusIconMini = () =>
     </g>
   </SVGIcon>
 
+export const ReplyIcon = () =>
+  <SVGIcon className="ReplyIcon">
+    <g>
+      <path d="M17,14.2c0-2.3-1.9-4.2-4.2-4.2H3" />
+    </g>
+    <g>
+      <polyline points="7,14 3,10 7,6 " />
+    </g>
+  </SVGIcon>
+
+export const RepostIcon = () =>
+  <SVGIcon className="RepostIcon">
+    <g className="svg-stroke-round">
+      <path d="M15.2,6.7H5 c-0.5,0-1,0.4-1,1V10" />
+      <path d="M4.8,14.4H15 c0.6,0,1-0.4,1-1v-2.3" />
+    </g>
+    <g className="svg-stroke-round">
+      <polyline points="13.3,4 16,6.7 13.2,9.5" />
+      <polyline points="6.7,17.1 4,14.4 6.8,11.7" />
+    </g>
+  </SVGIcon>
+
 export const RequestIcon = () =>
   <SVGIcon className="RequestIcon">
     <g>
@@ -250,7 +289,7 @@ export const ShareIcon = () =>
     </g>
     <g>
       <line x1="10" y1="2" x2="10" y2="12" />
-      <polyline points="7.2,4.7 10,2 12.8,4.8" />
+      <polyline className="svg-stroke-round" points="7.2,4.7 10,2 12.8,4.8" />
     </g>
   </SVGIcon>
 
@@ -258,6 +297,17 @@ export const SuccessIcon = () =>
   <SVGIcon className="SuccessIcon">
     <g>
       <polyline points="4.8,10.8 9.9,17 17.2,5" />
+    </g>
+  </SVGIcon>
+
+export const XBoxIcon = () =>
+  <SVGIcon className="XBoxIcon">
+    <g>
+      <rect x="3.5" y="3.5" width="12" height="12" />
+    </g>
+    <g>
+      <line x1="12" y1="7" x2="7" y2="12" />
+      <line x1="12" y1="12" x2="7" y2="7" />
     </g>
   </SVGIcon>
