+++ src/components/editor/RepostBlock.js
@@ -3,7 +3,7 @@
 import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
 import { LockIcon } from './EditorIcons'
-import { RepostIcon } from '../posts/PostIcons'
+import { RepostIcon } from '../assets/Icons'
 
 function mapStateToProps({ profile: currentUser }) {
   return { currentUser }
