--- src/components/posts/PostIcons.css
@@ -1,48 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-/* stylelint-disable selector-max-compound-selectors, selector-max-specificity, selector-no-qualifying-type */
-
-.Post .RepostIcon > g,
-.Post .ShareIcon > polyline {
-  stroke-linejoin: round;
-}
-
-.Post .PencilIcon > g,
-.Post .XBoxIcon > g {
-  stroke-linecap: square;
-  transition-delay: 0.1s, 0s, 0s;
-}
-
-.RepostHeader .RepostIcon > g {
-  stroke: #535353;
-}
-
-/* // ------------------------------------- */
-/* // PostTools */
-
-.no-touch .LoveTool > button:hover > .HeartIcon > g {
-  fill: #000;
-}
-
-.LoveTool > button.isActive > .HeartIcon > g {
-  fill: #f00;
-  stroke: #f00;
-}
-
-.no-touch .LoveTool > button.isActive:hover > .HeartIcon > g {
-  fill: #f99;
-  stroke: #f99;
-}
-
-.no-touch .ShareTool > button:hover > .ShareIcon > g:nth-of-type(2) {
-  transform: translate3d(0, -1px, 0);
-}
-
-.no-touch .DeleteTool > button:hover > .XBoxIcon > g:nth-of-type(2) {
-  transform: rotate(90deg);
-}
-
-/* // ------------------------------------- */
-
-/* stylelint-enable selector-max-compound-selectors, selector-max-specificity, selector-no-qualifying-type */
-
