--- src/components/posts/PostIcons.js
@@ -1,65 +1,0 @@
-/* eslint-disable max-len */
-import React from 'react'
-import { SVGIcon } from '../svg/SVGComponents'
-
-export const HeartIcon = () =>
-  <SVGIcon className="HeartIcon">
-    <g>
-      <path d="M10,7.4c0-1.8,1.5-3.2,3.3-3.2s3.3,1.4,3.3,3.2c0,4.5-6.5,8.4-6.5,8.4S3.5,12,3.5,7.4c0-1.8,1.5-3.2,3.3-3.2S10,5.6,10,7.4z" />
-    </g>
-  </SVGIcon>
-
-export const PencilIcon = () =>
-  <SVGIcon className="PencilIcon">
-    <g>
-      <polygon points="12.6,4 16,7.3 7.4,15.9 4,15.9 4,12.5" />
-    </g>
-    <g>
-      <line x1="10.2" y1="6.4" x2="13.6" y2="9.7" />
-    </g>
-  </SVGIcon>
-
-export const ReplyIcon = () =>
-  <SVGIcon className="ReplyIcon">
-    <g>
-      <path d="M17,14.2c0-2.3-1.9-4.2-4.2-4.2H3" />
-    </g>
-    <g>
-      <polyline points="7,14 3,10 7,6 " />
-    </g>
-  </SVGIcon>
-
-export const RepostIcon = () =>
-  <SVGIcon className="RepostIcon">
-    <g>
-      <path d="M15.2,6.7H5 c-0.5,0-1,0.4-1,1V10" />
-      <path d="M4.8,14.4H15 c0.6,0,1-0.4,1-1v-2.3" />
-    </g>
-    <g>
-      <polyline points="13.3,4 16,6.7 13.2,9.5" />
-      <polyline points="6.7,17.1 4,14.4 6.8,11.7" />
-    </g>
-  </SVGIcon>
-
-export const ShareIcon = () =>
-  <SVGIcon className="ShareIcon">
-    <g>
-      <polyline points="7.8,7.3 5,7.3 5,17.3 15,17.3 15,7.3 12.2,7.3" />
-    </g>
-    <g>
-      <line x1="10" y1="2" x2="10" y2="12" />
-      <polyline points="7.2,4.7 10,2 12.8,4.8" />
-    </g>
-  </SVGIcon>
-
-export const XBoxIcon = () =>
-  <SVGIcon className="XBoxIcon">
-    <g>
-      <rect x="3.5" y="3.5" width="12" height="12" />
-    </g>
-    <g>
-      <line x1="12" y1="7" x2="7" y2="12" />
-      <line x1="12" y1="12" x2="7" y2="7" />
-    </g>
-  </SVGIcon>
-
