+++ src/main.css
@@ -79,7 +79,6 @@
 @import "components/posts/Post.css";
 @import "components/posts/PostBody.css";
 @import "components/posts/PostHeader.css";
-@import "components/posts/PostIcons.css";
 @import "components/posts/PostTools.css";
 
 @import "components/relationships/BlockMuteButton.css";
